Create upload_dataser
This PR includes a Python script to prepare the parking transaction dataset before uploading it to BigQuery.

Main changes:

Column renaming to remove spaces and problematic characters.
Conversion of date columns to the proper format
Clean CSV export
Please check if the date format and the final name of the exported file is correct.